requirements.txt
requests
beautifulsoup4
pandas
matplotlib
seaborn
numpy
scipy
scikit-learn
jupyter
notebook
sqlalchemy
google-cloud-bigquery